+++ src/Bot.php
@@ -20,6 +20,7 @@
 namespace Baidu\Duer\Botsdk;
 
 use \Baidu\Apm\BotMonitorsdk;
+use \Curl\Curl;
 
 abstract class Bot{
 	const DEFAULT_EVENT_NAME = '__default__';
@@ -580,6 +581,20 @@ abstract class Bot{
     public function setFallBack(){
         $this->response->setFallBack();
     }
+    
+    /**
+     * @desc 表示directives中指令顺序随机
+     **/
+    public function setAutoDirectivesArrangement(){
+        $this->response->setAutoDirectivesArrangement();
+    }
+
+    /**
+     * @desc 表示directives中指令保持相对顺序不变 (directives中指令可能会被过滤)
+     **/
+    public function setStrictDirectivesArrangement(){
+        $this->response->setStrictDirectivesArrangement();
+    }
 
     /**
      * @desc 判断设备是否支持Interface
@@ -626,6 +641,162 @@ abstract class Bot{
     public function defaultEvent($event = null){
         $this->waitAnswer();
         $this->setExpectSpeech(false);
+    }
+
+    /**
+     * 技能所期待的用户回复，技能将该信息反馈给DuerOS，有助于DuerOS在语音识别以及识别纠错时向该信息提权。
+     * 普通文本
+     * @param string $text 普通文本内容类型回复表达的回复内容。
+     */
+    public function addExpectTextResponse($text){
+        $this->response->addExpectTextResponse($text);    
+    }
+
+    /**
+     * 技能所期待的用户回复，技能将该信息反馈给DuerOS，有助于DuerOS在语音识别以及识别纠错时向该信息提权。
+     * 槽位类型
+     * @param string $slot 槽位类型回复表达的槽位名称。
+     */
+    public function addExpectSlotResponse($slot){
+        $this->response->addExpectTextResponse($slot);    
+    }
+
+    /**
+     * 获取apiAccessToken
+     * @return string
+     */
+    public function getApiAccessToken(){
+        return $this->request->getApiAccessToken(); 
+    }
+
+    /**
+     * 获取apiEndPoint
+     * @return string
+     */
+    public function getApiEndPoint(){
+        return $this->request->getApiEndPoint(); 
+    }
+
+    /**
+     * curl封装
+     * @param array $opts
+     * @return mixed
+     */
+    public function curl($opts){
+        $apiAccessToken = $this->getApiAccessToken();
+        $apiEndPoint = $this->getApiEndPoint();
+        $headers = array(
+            'Authorization' => 'bearer ' . $apiAccessToken 
+        );
+        $defaultOpts = [
+            'url' => '',
+            'method' => 'get',
+            'timeout' => 1,
+            'uri' => $apiEndPoint,
+            'path' => '',
+            'data' => [],
+            'headers' => $headers,
+        ];
+        $opts = array_merge($defaultOpts, $opts);
+
+        $url = $opts['url'] ? : $opts['uri'] . $opts['path'];
+        $curl = new Curl();
+        $curl->setHeaders($opts['headers']);
+        $curl->setRetry(1);
+        $curl->setTimeout($opts['timeout']);
+        if($opts['method'] == 'get'){
+            $curl->get($url, $opts['data']);
+        }else if($opts['method'] == 'post'){
+            $curl->post($url, $opts['data']);
+        }
+
+        if ($curl->error) {
+            return false;
+        } else {
+            return $curl->response;
+        }
+    }
+
+    /**
+     * 获取用户百度账号信息
+     * 需要用户同意该权限才可以获取到信息，参考https://dueros.baidu.com/didp/doc/dueros-bot-platform/dbp-user-info/request-customer-information-api_markdown
+     * @return mixed
+     */
+    public function getUserProfile(){
+        $opts = array(
+            'path' => '/saiya/v1/user/profile' 
+        );
+        return $this->curl($opts);
+    }
+
+    /**
+     * 获取用户录音数据
+     * 需要用户同意该权限才可以获取到信息，参考https://dueros.baidu.com/didp/doc/dueros-bot-platform/dbp-user-info/request-customer-information-api_markdown
+     * @param string $audioToken RecordSpeechFinished事件中的audioToken
+     * @return mixed
+     */
+    public function getRecordSpeech($audioToken){
+        $opts = array(
+            'path' => '/saiya/v1/user/record/speech',
+            'data' => array(
+                'audioToken' => $audioToken,
+            ) 
+        );
+        return $this->curl($opts);
+    }
+
+    /**
+     * 获取用户地理位置信息
+     * 需要用户同意该权限才可以获取到信息，参考https://dueros.baidu.com/didp/doc/dueros-bot-platform/dbp-user-info/request-customer-information-api_markdown
+     * @return mixed
+     */
+    public function getDeviceLocation(){
+        $opts = array(
+            'path' => '/saiya/v1/devices/location',
+        );
+        return $this->curl($opts);
+    }
+
+    /**
+     * 调用智能家居打印机服务
+     * 需要用户同意该权限才可以获取到信息，参考https://dueros.baidu.com/didp/doc/dueros-bot-platform/dbp-user-info/request-customer-information-api_markdown
+     * @param array $data 
+     * @return mixed
+     */
+    public function callSmarthomePrinter($data){
+        $apiAccessToken = $this->getApiAccessToken();
+        $headers = array(
+            'Authorization' => 'bearer ' . $apiAccessToken,
+            'Content-Type' => 'application/json' 
+        );
+        $opts = array(
+            'path' => '/saiya/v1/smarthome/printer',
+            'method' => 'post',
+            'data' => $data,
+            'headers' => $headers
+        );
+        return $this->curl($opts);
+    }
+
+    /**
+     * 小度音响app通知接口
+     * 需要用户同意该权限才可以获取到信息，参考https://dueros.baidu.com/didp/doc/dueros-bot-platform/dbp-user-info/request-customer-information-api_markdown
+     * @param array $data 
+     * @return mixed
+     */
+    public function sendMateappNotification($data){
+        $apiAccessToken = $this->getApiAccessToken();
+        $headers = array(
+            'Authorization' => 'bearer ' . $apiAccessToken,
+            'Content-Type' => 'application/json' 
+        );
+        $opts = array(
+            'path' => '/saiya/v1/mateapp/notification',
+            'method' => 'post',
+            'data' => $data,
+            'headers' => $headers
+        );
+        return $this->curl($opts);
     }
 
 }